+++ packages/ir-sdk/fern/apis/ir-types-latest/changelog/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v44] - 2024-05-10
+
+- Improvement: Support stream and server-sent event response examples.
+
 ## [v43] - 2024-05-08
 
 - Improvement: Support custom status codes for success respones.